Readd access right

+++ src/Controller/ProjectController.php
@@ -37,6 +37,8 @@ class ProjectController extends AbstractController
     #[Route('/project/{id}', name: 'app_project_show')]
     public function show(Project $project): Response
     {
+        $this->denyAccessUnlessGranted('access', $project);
+
         return $this->render('project/show.html.twig', [
             'project' => $project,
         ]);
@@ -45,6 +47,8 @@ class ProjectController extends AbstractController
     #[Route('/project/{id}/edit', name: 'app_project_edit')]
     public function edit(Request $request, Project $project, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted('access', $project);
+
         $form = $this->createForm(ProjectType::class, $project);
         $form->handleRequest($request);
 
@@ -62,6 +66,8 @@ class ProjectController extends AbstractController
     #[Route('/project/{id}/delete', name: 'app_project_delete')]
     public function delete(Project $project, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted('access', $project);
+
         $entityManager->remove($project);
         $entityManager->flush();
